[ClockingIT @ New version of acts_as_ferret with DRb support]
I had to upgrade acts_as_ferret to a new version today, which
has DRb support, as I was experiencing quite a bit of index
corruption due to a high amount of concurrent modification.

This seems to have solved it, but it might require an upgrade
of ferret on your end if you experience problems.