[ClockingIT @ Move files out of database and into a storage directory]
Quite a big patch, this, and it's the beginnings of moving the files
stored in the database out into the filesystem instead.

Later on I'll drop the existing data in the database, but for now
old binaries will be stored both places, but new ones are only in
the filesystem.

'rake migrate' required