added new time format, removed pause button
The new time format adds the possibiliy to display time without leading
zeros. (9:15 instead of 09:15). I had to modify datejs a bit to comply
with rails time formats though.
I removed the code in the views that permitted ppl to pause time because
the pause time cause too much trouble for the accountant. We need to
discuss if we permanently remove all code for the paused time (from
controllers models etc) or not