finalised default_project, modified options_for_projects stuff
A user can now select a default project in his preferences. This project
will be selected by default when he creates a new task.

options_for_projects was moved from the notice_groups helper to the
application helper so that it can be used everywhere where a project
menu needs to be used.